get user logged in status in layout and pass it down to navbar
